agent/sub_agents/logic_check/LogicCheck.py
import os
from google.adk.agents import Agent, ParallelAgent
from google.adk.tools.agent_tool import AgentTool
from google.genai.types import GenerateContentConfig
from google.adk.agents import LlmAgent
from google.adk.models.lite_llm import LiteLlm

_prompt = """
你是一个负责检查文案逻辑的agent，请检查输入的文案是否符合以下逻辑标准：

1. 内容结构逻辑：
   - 段落之间是否有清晰的逻辑关系和过渡
   - 论述是否存在前后矛盾
   - 各部分内容是否围绕核心主题展开
   - 是否存在逻辑跳跃或断层

2. 因果关系逻辑：
   - 因果关系是否明确且合理
   - 论据是否充分支持论点
   - 结论是否基于充分的论述
   - 是否存在无根据的推断

3. 表达逻辑：
   - 语言表达是否清晰准确
   - 专业术语使用是否恰当
   - 数据引用是否准确且有出处
   - 比喻和类比是否恰当

4. 目标导向逻辑：
   - 内容是否符合目标受众的认知水平
   - 价值主张是否符合用户需求
   - 行动号召是否与内容匹配
   - 是否存在与品牌调性不符的表达

如果发现逻辑问题，请：
1. 指出具体的逻辑问题所在
2. 分析问题产生的原因
3. 提供修改建议
4. 给出修改后的示例

评估维度：
- 逻辑严密性：内容是否存在逻辑漏洞
- 可理解性：受众是否容易理解
- 说服力：论述是否具有说服力
- 连贯性：文案整体是否连贯流畅

如果完全符合逻辑标准，请返回"文案逻辑合理，结构清晰"并给出积极的反馈。
"""

qwen_model = LiteLlm(
    model="openai/Qwen/Qwen3-235B-A22B",
    api_base=os.environ.get("OPENAI_API_BASE"),
    api_key=os.environ.get("OPENAI_API_KEY"),
    extra_body={"enable_thinking": False, "stream": False},
    stream=False,
)

logic_check = LlmAgent(
    model=qwen_model,
    name="logic_check", 
    description="""一个负责检查文案逻辑的agent""",
    instruction=_prompt,
) 